+++ test/dummy/test/integration/default_layout_assets_test.rb
@@ -28,7 +28,9 @@ class DefaultLayoutAssetsTest < ActionDispatch::IntegrationTest
     assert_includes response.body, "/assets/tailwind"
     assert_includes response.body, "@hotwired/turbo-rails"
     refute_includes response.body, "dummy_page_nav"
-    assert_includes response.body, "Sign out"
+    refute_includes response.body, "Sign out"
+    refute_includes response.body, 'href="/users/sign_out"'
+    refute_includes response.body, "/recording_studio_root_switchable/v1/root_switch"
     assert_select "[aria-label='Go back']", count: 1
     assert_select "a[aria-label='Close']", count: 1
   end